Refactoring project with conversion-tracer-bullet module
