+++ conversiontracerbullet/src/main/java/org/stanford/MarcToXML.java
@@ -1,4 +1,4 @@
-package main;
+package org.stanford;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
