+++ conversiontracerbullet/src/test/resources/sample_marc.mrc
@@ -1,0 +1,1 @@
+01373cam a2200433 a 45000010008000000010012000080030006000200080041000260100016000670200015000830200018000980200015001160200018001310200015001490200018001640350020001820350019002020350022002210350023002430400027002660430012002930450009003050820008003141000045003222400019003672450051003862600053004373000017004905000050005075050050005575960006006076000061006136000060006747000044007347000031007789100099008099830019009089180012009274241894it 9932055XSIRSI990920m19999999it a     b    000 0bita d  a   00275292  a8887114404  a9788887114409  a8887114587  a9788887114584  a8884980151  a9788884980151  aILCGHZ4030041-B  a(ICU)hz4030041  a(OCoLC-M)42416090  a(OCoLC-I)275650394  aC3LcC3LdCGUdCStRLIN  ae-it---  ax0x414a8561 aDe Luca, Giuseppe,d1898-1962.=^A10647510aCorrespondence10aCarteggio /cGiuseppe De Luca, Fausto Minelli.  aRoma :bEdizioni di storia e letteratura,c1999-  av. ;c26 cm.  aLetters, mostly published for the first time.1 a1. 1930-1934 -- 2. 1935-1939 -- 3. 1940-1946.  a110aDe Luca, Giuseppe,d1898-1962vCorrespondence.=^A10647510aMinelli, Fausto,d1891-1974vCorrespondence.=^A15253091 aMinelli, Fausto,d1891-1974.=^A15253091 aRoncalli, Marco.=^A902247  aSupplied on B.O. c3NLC010 dCAX040B eStanford University Libraries fCASALI gCasalini Libri  a36105021876367  a4241894
