--- src/test/MarcToXMLTest.java
@@ -1,19 +1,0 @@
-package test;
-
-import org.junit.Test;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-
-public class MarcToXMLTest {
-
-    private static final Logger log = LogManager.getLogger();
-
-    @Test
-    public void main() throws Exception {
-        System.out.println("MarctoXML Test");
-        log.info("Testing MarcToXML");
-
-    }
-
-}
