--- src/test/resources/log4j2-test.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- This configuration can be overridden by placing a file log4j2-test.xml in the same directory (src). -->
-<!-- Status = level of internal log4j events to log to console. Set status to "trace" to troubleshoot log4j2. -->
-<Configuration status="info">
-    <Appenders>
-        <Console name="Console" target="SYSTEM_OUT">
-            <PatternLayout pattern="%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} line %L - %msg%n"/>
-            <!-- <PatternLayout pattern="%d{yyyy-MM-dd HH:mm:ss.SSS} - %msg%n"/> -->
-        </Console>
-        <!-- Note: unless this is commented out, the file will be created, even if
-          it is empty. To write to the file comment in the AppenderRef. -->
-        <RollingFile name="RollingFileDev" fileName="log/ld4p-tracer-bullets.log"
-                     filePattern="log/$${date:yyyy-MM-dd}/bib2lod-%d{yyyy-MM-dd-HHmmss}.log">
-            <PatternLayout pattern="%d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %-5level %logger{36} line %L - %msg%n"/>
-            <Policies>
-                <OnStartupTriggeringPolicy />
-            </Policies>
-        </RollingFile>
-        <!--
-        <RollingFile name="RollingFile" fileName="log/ld4p-tracer-bullets.log"
-                     filePattern="log/$${date:yyyy-MM-dd}/bib2lod-%d{yyyy-MM-dd-HHmmss}.log">
-            <PatternLayout pattern="%d{yyyy-MM-dd HH:mm:ss.SSS} - %msg%n"/>
-            <Policies>
-                <OnStartupTriggeringPolicy />
-            </Policies>
-        </RollingFile>
-        -->
-    </Appenders>
-    <Loggers>
-        <Root level="debug">
-            <AppenderRef ref="Console"/>
-            <!-- <AppenderRef ref="RollingFileDev"/> -->
-        </Root>
-        <!-- Set additivity to false since the Root logger also writes to the console - prevents
-                repeat logging -->
-        <Logger name="package org.ld4l.bib2lod.SimpleManager" additivity="false" level="debug">
-            <AppenderRef ref="Console"/>
-        </Logger>
-    </Loggers>
-</Configuration>
